chore: Adding the ability to decide how to merge states
A merger function can be passed through the UohStore constructor or individually through the setState method. This is done to allow for users to control how to merge current state with a new state. Default merger function is set to merge using the Object assign function, to prevent any breaking changes.